+++ libs/qwt/src/qwt_plot_intervalcurve.h
@@ -106,7 +106,7 @@ class QWT_EXPORT QwtPlotIntervalCurve
 
     virtual void drawSeries( QPainter*,
         const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-        const QRectF& canvasRect, size_t from, size_t to ) const QWT_OVERRIDE;
+        const QRectF& canvasRect, int from, int to ) const QWT_OVERRIDE;
 
     virtual QRectF boundingRect() const QWT_OVERRIDE;
